+++ src/main.rs
@@ -2,7 +2,7 @@ use clap::{Parser, Subcommand};
 mod alias;
 
 #[derive(Parser)]
-#[command(name = "cliq", version, about , author, long_about = None)]
+#[command(name = "cliq", version, about, author, long_about = None)]
 struct Cliq {
     #[command(subcommand)]
     command: Commands,
@@ -10,6 +10,8 @@ struct Cliq {
 
 #[derive(Subcommand, Debug)]
 enum Commands {
+    #[command(about = "get all aliases from cliq.toml")]
+    List,
     // to handle aliases in toml file
     #[command(external_subcommand)]
     #[allow(dead_code)]
@@ -20,6 +22,10 @@ fn main() {
     let cliq = Cliq::parse();
 
     match &cliq.command {
+        Commands::List => {
+            let parsed_toml = alias::links();
+            print!("{}", parsed_toml);
+        }
         Commands::Options(args) => {
             let input = args.join(" ");
             let url = alias::link(input);