+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### Added
 - support for linux
+- `list` command to get all aliases
 - clap cli parser
 
 ## [0.1.0] - 2024-05-24